src/requirements.txt
asttokens==2.0.5
colorama==0.4.4
executing==0.6.0
icecream==2.1.1
Pygments==2.9.0
PyYAML==5.4.1
six==1.16.0